fix: address GPT-5.6 review — feature gate, sync facade, archive path, CI
