+++ .github/workflows/ci.yml
@@ -60,3 +60,21 @@ jobs:
 
       - name: Build
         run: cargo build --release
+
+  msrv:
+    name: MSRV 1.85
+    runs-on: ubuntu-latest
+    permissions:
+      contents: read
+
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0 # v7.0.0
+        with:
+          persist-credentials: false
+
+      - name: Install Rust 1.85
+        run: rustup toolchain install 1.85 --profile minimal --no-self-update
+
+      - name: Build with MSRV
+        run: cargo +1.85 build --release
